mcp_server/requirements.txt
# MCP Server dependencies
fastmcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
